implemented the discharge of Drone.cpp

+++ src/control_center/ControlCenter.cpp
@@ -21,6 +21,9 @@
 
 // TODO: gestire chiusura di tutti i while true
 // TODO: a ogni cambio di stato del drone fare l'update nel db
+// TODO: gestire la batteria del drone (Drone.cpp)
+// TODO: Mandare posizione CC ai droni
+// TODO: implementare che il drone si scarica mentre lavora
 
 
 
@@ -213,8 +216,8 @@ void ControlCenter::initDrones() {
         Redis::Message message = get<1>(response);
         DroneData droneData = DroneData();
         droneData.id = stoi(message["id"]);
-        droneData.x = stof(message["x"]);
-        droneData.y = stof(message["y"]);
+        droneData.x = stoi(message["x"]);
+        droneData.y = stoi(message["y"]);
         droneData.battery = stof(message["battery"]);
         droneData.state = to_state(message["state"]);
